# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/portal_users:
    get:
      tags:
        - PortalUsers
      summary: List PortalUsers
      description: Get a paginated list of PortalUsers with optional filtering and sorting
      operationId: listPortalUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalUserList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PortalUsers
      summary: Create PortalUser
      description: Create a new PortalUser
      operationId: createPortalUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePortalUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalUser'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PortalUser:
      type: object
      required:
        - id
        - email
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          maxLength: 320
        display_name:
          nullable: true
          type: string
          maxLength: 120
        status:
          type: string
        phone:
          nullable: true
          type: string
          maxLength: 40
        street:
          nullable: true
          type: string
          maxLength: 160
        street2:
          nullable: true
          type: string
          maxLength: 160
        city:
          nullable: true
          type: string
          maxLength: 80
        state_id:
          nullable: true
          type: string
          maxLength: 80
        country_id:
          nullable: true
          type: string
          maxLength: 8
        zip:
          nullable: true
          type: string
          maxLength: 24
        vat:
          nullable: true
          type: string
          maxLength: 40
        company_name:
          nullable: true
          type: string
          maxLength: 120
        sapiens_user_id:
          nullable: true
          type: string
          format: uuid
        revoked_at:
          nullable: true
          type: string
          format: date-time
        revocation_reason:
          nullable: true
          type: string
          maxLength: 120
        archived_at:
          nullable: true
          type: string
          format: date-time
        last_login_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    CreatePortalUserRequest:
      type: object
      required:
        - email
        - status
        - metadata
      properties:
        email:
          type: string
          maxLength: 320
        display_name:
          nullable: true
          type: string
          maxLength: 120
        status:
          type: string
        phone:
          nullable: true
          type: string
          maxLength: 40
        street:
          nullable: true
          type: string
          maxLength: 160
        street2:
          nullable: true
          type: string
          maxLength: 160
        city:
          nullable: true
          type: string
          maxLength: 80
        state_id:
          nullable: true
          type: string
          maxLength: 80
        country_id:
          nullable: true
          type: string
          maxLength: 8
        zip:
          nullable: true
          type: string
          maxLength: 24
        vat:
          nullable: true
          type: string
          maxLength: 40
        company_name:
          nullable: true
          type: string
          maxLength: 120
        sapiens_user_id:
          nullable: true
          type: string
          format: uuid
        revoked_at:
          nullable: true
          type: string
          format: date-time
        revocation_reason:
          nullable: true
          type: string
          maxLength: 120
        archived_at:
          nullable: true
          type: string
          format: date-time
        last_login_at:
          nullable: true
          type: string
          format: date-time
        metadata:
          type: object
...
    PortalUserList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PortalUser'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []