# src/presentation/http/portal_user_handler.rs
//! PortalUser REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{DateTime, Utc};

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{PortalUserService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreatePortalUserDto, UpdatePortalUserDto, PatchPortalUserDto, PortalUserResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum PortalUserError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for PortalUserError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for PortalUserError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "PORTALUSER_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "PORTALUSER_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PORTALUSER_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "PORTALUSER_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for PortalUser.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /portal_users | List with pagination |
/// | POST | /portal_users | Create new |
/// | GET | /portal_users/:id | Get by ID |
/// | PUT | /portal_users/:id | Full update |
/// | PATCH | /portal_users/:id | Partial update |
/// | DELETE | /portal_users/:id | Soft delete |
/// | POST | /portal_users/bulk | Bulk create |
/// | POST | /portal_users/upsert | Upsert |
/// | GET | /portal_users/trash | List deleted |
/// | POST | /portal_users/:id/restore | Restore |
/// | DELETE | /portal_users/empty | Empty trash |
/// | GET | /portal_users/:id/deleted | Get deleted by ID |
/// | DELETE | /portal_users/trash/:id | Permanent delete from trash |
/// | GET | /portal_users/count | Count active entities |
/// | GET | /portal_users/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(PortalUserService::with_repository(repository));
/// let router = create_portal_user_routes(service);
/// ```
pub fn create_portal_user_routes(service: Arc<PortalUserService>) -> Router {
    BackboneCrudHandler::<PortalUserService, PortalUser, CreatePortalUserDto, UpdatePortalUserDto, PortalUserResponseDto>::routes(
        service,
        "/portal_users",
    )
}

/// Create Axum router with only the read (GET) endpoints for PortalUser.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_portal_user_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_portal_user_read_routes(service: Arc<PortalUserService>) -> Router {
    BackboneCrudHandler::<PortalUserService, PortalUser, CreatePortalUserDto, UpdatePortalUserDto, PortalUserResponseDto>::read_routes(
        service,
        "/portal_users",
    )
}

/// Create Axum router with only the write (mutation) endpoints for PortalUser.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
///
/// # This is unguarded generic CRUD, not a validated write path
///
/// These are plain create/update/patch/delete mutations over the entity row —
/// they bypass all business invariants. If the module exposes a validated write
/// service (e.g. a command router over its domain engine), serve THAT instead
/// for any mutation that must respect domain rules.
pub fn create_portal_user_write_routes(service: Arc<PortalUserService>) -> Router {
    BackboneCrudHandler::<PortalUserService, PortalUser, CreatePortalUserDto, UpdatePortalUserDto, PortalUserResponseDto>::write_routes(
        service,
        "/portal_users",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_portal_user_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<PortalUserService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_portal_user_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}